refactor: Move Enums to _core.commons package
