+++ src/test/java/com/bungaebowling/server/user/controller/UserControllerTest.java
@@ -1,9 +1,9 @@
 package com.bungaebowling.server.user.controller;
 
 import com.amazonaws.services.s3.AmazonS3;
-import com.bungaebowling.server.ApiTag;
 import com.bungaebowling.server.ControllerTestConfig;
-import com.bungaebowling.server.GeneralApiResponseSchema;
+import com.bungaebowling.server._core.commons.ApiTag;
+import com.bungaebowling.server._core.commons.GeneralApiResponseSchema;
 import com.bungaebowling.server._core.errors.exception.ErrorCode;
 import com.bungaebowling.server._core.security.JwtProvider;
 import com.bungaebowling.server.user.Role;
